CMA-62 [refactor] : 커밋후 생긴 삭제 및 수정사항
