--- src/main/java/com/team/comma/ValueupApplication.java
@@ -1,13 +1,0 @@
-package com.team.comma;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ValueupApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ValueupApplication.class, args);
-	}
-
-}
