+++ src/main/java/com/team/comma/repository/UserRepository.java
@@ -2,7 +2,7 @@ package com.team.comma.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team.comma.entity.User;
+import com.team.comma.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
