+++ src/main/java/com/team/comma/util/security/CustomUserDetailService.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.team.comma.entity.User;
+import com.team.comma.domain.User;
 import com.team.comma.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
