--- src/main/resources/application-oauth.properties
@@ -1,35 +1,0 @@
-# Kakao
-spring.security.oauth2.client.registration.kakao.client-id=54a1a647665a4c69080a68825c0982a1
-spring.security.oauth2.client.registration.kakao.client-secret=unknown
-spring.security.oauth2.client.registration.kakao.redirect-uri=http://localhost:3000/oauth/login
-spring.security.oauth2.client.registration.kakao.authorization-grant-type=authorization_code
-spring.security.oauth2.client.registration.kakao.scope=account_email
-spring.security.oauth2.client.registration.kakao.client-name=kakao
-spring.security.oauth2.client.registration.kakao.client-authentication-method=POST
-
-# Google
-spring.security.oauth2.client.registration.google.client-id=505351076247-r4tihp3g0d6hdg9vrhqlk8bdfsnr93ck.apps.googleusercontent.com
-spring.security.oauth2.client.registration.google.client-secret=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
-spring.security.oauth2.client.registration.google.redirect-uri=http://localhost:3000/oauth/login
-spring.security.oauth2.client.registration.google.scope=email
-
-# Naver
-spring.security.oauth2.client.registration.naver.client-id=cRB2fcyuQl_QRISKVV0a
-spring.security.oauth2.client.registration.naver.client-secret=iVKKbgoepS
-spring.security.oauth2.client.registration.naver.redirect-uri=http://localhost:3000/oauth/login
-spring.security.oauth2.client.registration.naver.authorization-grant-type=authorization_code
-spring.security.oauth2.client.registration.naver.client-name=naver
-
-# provider
-
-# Kakao
-spring.security.oauth2.client.provider.kakao.authorization-uri=https://kauth.kakao.com/oauth/authorize
-spring.security.oauth2.client.provider.kakao.token-uri=https://kauth.kakao.com/oauth/token
-spring.security.oauth2.client.provider.kakao.user-info-uri=https://kapi.kakao.com/v2/user/me
-spring.security.oauth2.client.provider.kakao.user-name-attribute=id
-
-# Naver
-spring.security.oauth2.client.provider.naver.authorization-uri=https://nid.naver.com/oauth2.0/authorize
-spring.security.oauth2.client.provider.naver.token-uri=https://nid.naver.com/oauth2.0/token
-spring.security.oauth2.client.provider.naver.user-info-uri=https://openapi.naver.com/v1/nid/me
-spring.security.oauth2.client.provider.naver.user-name-attribute=response
