+++ src/test/java/com/team/comma/repository/RefreshTokenRepositoryTest.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import com.team.comma.entity.RefreshToken;
+import com.team.comma.domain.RefreshToken;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE) // EmbeddedDatabase 가 아닌 Mysql 에 테스트 사용됨
