+++ src/test/java/com/team/comma/service/JwtServiceTest.java
@@ -18,8 +18,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.team.comma.dto.MessageResponse;
-import com.team.comma.entity.RefreshToken;
-import com.team.comma.entity.Token;
+import com.team.comma.domain.RefreshToken;
+import com.team.comma.domain.Token;
 import com.team.comma.exception.FalsifyTokenException;
 import com.team.comma.repository.RefreshTokenRepository;
 import com.team.comma.util.security.JwtTokenProvider;
